http://issues.apache.org/activemq/browse/AMQCPP-30
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@505578 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/concurrent/CountDownLatch.cpp
@@ -56,7 +56,7 @@ void CountDownLatch::await() throw ( cms::CMSException ) {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-void CountDownLatch::await( long long timeOut ) throw ( cms::CMSException ) {
+void CountDownLatch::await( unsigned long timeOut ) throw ( cms::CMSException ) {
     try {
 
         synchronized( &mutex ) {
