+++ activemq-cpp/src/main/activemq/concurrent/CountDownLatch.h
@@ -57,7 +57,7 @@ namespace concurrent{
          * Waits for the Count to hit zero, or a timeout.
          * @param timeOut - time in milliseconds to wait.
          */
-        virtual void await( long long timeOut )  throw ( cms::CMSException );
+        virtual void await( unsigned long timeOut )  throw ( cms::CMSException );
 
         /**
          * Counts down the latch, releasing all waiting threads when
